+++ VoucherServiceBL/Model/VoucherUpdateReq.cs
@@ -7,7 +7,7 @@ namespace VoucherServiceBL.Model
     {
         public string Status { get; set; }
         public DateTime ExpiryDate { get; set; }
-        public BigInteger GiftAmount { get; set; }
+        public long GiftAmount { get; set; }
         public string Code { get; set; }
     }
 }
